+++ src/main/java/next/controller/users/ProfileController.java
@@ -1,7 +1,7 @@
 package next.controller.users;
 
 import core.db.DataBase;
-import next.controller.Controller;
+import core.mvc.Controller;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
